Move API validation
